Update the comic UI to reflect the settings.
In this patch we need to do somewhat of an ugly hack to be able to update the UI properly. It seems, if we only try to update the ObservableCollection with a new LINQ execution, the collection is not really updated. So we end up creating a extra copy with the new items from the DB and iterating these items to the ObservableCollection that is in the UI.